+++ preframr_tokens/__init__.py
@@ -24,7 +24,14 @@ from preframr_tokens.macros.roles import (
 from preframr_tokens.audit_primitives import (
     detect_tail_cycle,
     distinct_n,
+    op_atom_profile,
+    register_state,
     tier_accuracy,
+    trajectory_coverage,
+)
+from preframr_tokens.tokenizer_config import (
+    default_tokenizer_args,
+    named_config,
 )
 from preframr_tokens.vocab_signature import CONTENT_TIER, VocabSignature
 from preframr_tokens.tier_classify import (
@@ -53,7 +60,7 @@ from preframr_tokens.macros.transform import (
     register,
 )
 from preframr_tokens.macros import validate_back_refs, validate_pattern_overlays
-from preframr_tokens.macros.slope_pass import SlopePass
+from preframr_tokens.macros.freq_trajectory_pass import FreqTrajectoryPass
 from preframr_tokens.macros.preset_pass import PresetPass
 from preframr_tokens.macros.per_reg_burst import PerRegBurstPass
 from preframr_tokens.macros.gate_slope_shift_pass import GateSlopeShiftPass
@@ -89,7 +96,7 @@ __all__ = [
     "PassBackedTransform",
     "RowExpandingTransform",
     "DistancePairSpec",
-    "SlopePass",
+    "FreqTrajectoryPass",
     "PresetPass",
     "PerRegBurstPass",
     "GateSlopeShiftPass",
@@ -116,6 +123,11 @@ __all__ = [
     "tier_accuracy",
     "detect_tail_cycle",
     "distinct_n",
+    "register_state",
+    "op_atom_profile",
+    "trajectory_coverage",
+    "default_tokenizer_args",
+    "named_config",
     "distance_pair_role",
     "slope_subreg_role",
     "frame_weight_role",